remove old menu page and only use the one with the orders from now on
