--- src/app/(layout-free)/event/[slug]/menu-n-orders/layout.tsx
@@ -1,9 +1,0 @@
-import SyncProvider from "@/provider/syncProvider";
-
-export default function Layout({ children }: { children: React.ReactNode }) {
-  return (
-    <div className="min-h-screen flex flex-col bg-gray-900">
-      <SyncProvider>{children}</SyncProvider>
-    </div>
-  );
-}
