--- src/app/(layout-free)/event/[slug]/menu-n-orders/menu-item.tsx
@@ -1,40 +1,0 @@
-import iconMap from "@/components/icon-map";
-import { XIcon } from "lucide-react";
-
-export default function MenuItem({
-  title,
-  shortTitle,
-  description,
-  originalTitle,
-}: {
-  title: string;
-  shortTitle: string;
-  description: string;
-  originalTitle?: string;
-}) {
-  const IconComponent = iconMap[originalTitle ?? shortTitle];
-  return (
-    <div className="flex flex-col cursor-pointer">
-      {IconComponent ? (
-        <IconComponent
-          width="8rem"
-          height="8rem"
-          style={{ fill: "white" }}
-          className="m-2 mx-auto"
-        />
-      ) : (
-        <XIcon
-          width="3rem"
-          height="3rem"
-          fill="text-black"
-          style={{ fill: "white" }}
-          className="m-2 mx-auto"
-        />
-      )}
-      <div className="text-center">
-        <h3 className="text-[2.3rem] font-semibold">{title}</h3>
-        <p className="text-[1.8rem] mt-2 max-w-sm">{description}</p>
-      </div>
-    </div>
-  );
-}
