+++ src/app/(layout-free)/event/[slug]/menu/layout.tsx
@@ -2,7 +2,7 @@ import SyncProvider from "@/provider/syncProvider";
 
 export default function Layout({ children }: { children: React.ReactNode }) {
   return (
-    <div className="min-h-screen flex flex-col">
+    <div className="min-h-screen flex flex-col bg-gray-900">
       <SyncProvider>{children}</SyncProvider>
     </div>
   );
