+++ src/app/(layout-free)/event/[slug]/menu/menu-item.tsx
@@ -13,9 +13,6 @@ export default function MenuItem({
   originalTitle?: string;
 }) {
   const IconComponent = iconMap[originalTitle ?? shortTitle];
-  if(!IconComponent){
-    console.debug(`No icon found for menu item: ${shortTitle}`);
-  }
   return (
     <div className="flex flex-col cursor-pointer">
       {IconComponent ? (
@@ -35,8 +32,8 @@ export default function MenuItem({
         />
       )}
       <div className="text-center">
-        <h3 className="font-semibold">{title}</h3>
-        <p className="text-sm mt-2">{description}</p>
+        <h3 className="text-[2.3rem] font-semibold">{title}</h3>
+        <p className="text-[1.8rem] mt-2 max-w-sm">{description}</p>
       </div>
     </div>
   );
